backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.9
python-dotenv==1.1.1
onnxruntime==1.22.1
pillow==11.3.0
opencv-python==4.12.0.88
numpy==2.2.6
requests==2.32.5
